Refactor BaseShader module a bit so it can scale when we add lights and shadows
